Moved styles into their own folder
