+++ src/components/editor_top_bar.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import editorStyles from './editor.module.css'
+import editorStyles from './styles/editor.module.css'
 
 export default ({title}) => (
   <div className={editorStyles.topBar}>{title}</div>
