+++ src/components/page.js
@@ -3,7 +3,7 @@ import React from 'react'
 import Head from './head'
 import Editor from './editor'
 import Sidebar from './sidebar'
-import pageStyles from './page.module.css'
+import pageStyles from './styles/page.module.css'
 
 export default ({children, title, description, path}) => (
   <div className={pageStyles.page}>
